feat: thunks taken out in a separate file
